gen-doc: remove dependency on libseccomp-dev
Move functions from pkg/gadgets/seccomp/gadget.go into separate files:

* syscalls.go: used by the gadget, need libseccomp
* syscalls_none.go: shim for gen-doc, without dependency on libseccomp
